ml_backend/prediction_service/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.24.0

# Database
google-cloud-bigquery==3.17.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Validation
pydantic==2.6.1

# Additional Utilities
python-jose==3.3.0
python-multipart==0.0.9

# Logging and Monitoring
loguru==0.7.2
dotenv==0.19.0